auth package doesnt import cmd
