+++ cmd/restorer/main.go
@@ -9,9 +9,9 @@ import (
 	"github.com/buildpack/imgutil/remote"
 
 	"github.com/buildpack/lifecycle"
+	"github.com/buildpack/lifecycle/auth"
 	"github.com/buildpack/lifecycle/cache"
 	"github.com/buildpack/lifecycle/cmd"
-	"github.com/buildpack/lifecycle/image/auth"
 )
 
 var (
@@ -77,7 +77,7 @@ func restore() error {
 	if cacheImageTag != "" {
 		origCacheImage, err := remote.NewImage(
 			cacheImageTag,
-			auth.DefaultEnvKeychain(),
+			auth.EnvKeychain(cmd.EnvRegistryAuth),
 			remote.FromBaseImage(cacheImageTag),
 		)
 		if err != nil {
@@ -86,7 +86,7 @@ func restore() error {
 
 		emptyImage, err := remote.NewImage(
 			cacheImageTag,
-			auth.DefaultEnvKeychain(),
+			auth.EnvKeychain(cmd.EnvRegistryAuth),
 			remote.WithPreviousImage(cacheImageTag),
 		)
 		if err != nil {
