+++ go.mod
@@ -3,7 +3,7 @@ module github.com/buildpack/lifecycle
 require (
 	github.com/BurntSushi/toml v0.3.1
 	github.com/apex/log v1.1.2-0.20190827100214-baa5455d1012
-	github.com/buildpack/imgutil v0.0.0-20191010153712-78959154ded1
+	github.com/buildpack/imgutil v0.0.0-20191021214009-9ad9206a7b17
 	github.com/docker/docker v1.4.2-0.20190924003213-a8608b5b67c7
 	github.com/docker/go-connections v0.4.0
 	github.com/golang/mock v1.3.1
@@ -17,7 +17,5 @@ require (
 	golang.org/x/net v0.0.0-20190724013045-ca1201d0de80 // indirect
 	google.golang.org/genproto v0.0.0-20190508193815-b515fa19cec8 // indirect
 )
-
-replace github.com/buildpack/imgutil => ../imgutil
 
 go 1.13
