--- image/auth/reference.go
@@ -1,20 +1,0 @@
-package auth
-
-import (
-	"github.com/google/go-containerregistry/pkg/authn"
-	"github.com/google/go-containerregistry/pkg/name"
-)
-
-func ReferenceForRepoName(keychain authn.Keychain, ref string) (name.Reference, authn.Authenticator, error) {
-	var auth authn.Authenticator
-	r, err := name.ParseReference(ref, name.WeakValidation)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	auth, err = keychain.Resolve(r.Context().Registry)
-	if err != nil {
-		return nil, nil, err
-	}
-	return r, auth, nil
-}
